Add a new virtio-net device to communicate with a DPDK poll mode driver
This patch adds a new virtio-net device called virtio-net-ipc.
The device can communicate with a DPDK poll mode driver called virtq
PMD.

Here is a command line example to use the device.
 -device virtio-net-ipc-pci,id=net0,mac=<macaddr>,bus=pci.0,addr=0x9,
         nid=0,socketpath=/tmp/virtq,cinterval=1 \
 -net none

'socketpath' is a path of unix domain socket to communicate with virtq
PMD. 'nid' is a unique identifier assigned to this virtio-net-ipc device.
If 2 devices use a same 'socketpath', 'nid' should be different.
'cinterval' is a value of time interval(sec) during reconnection.